requirements.txt
# Core dependencies
Flask==3.1.2
markdown2==2.5.4
EbookLib==0.19
reportlab==4.0.7

# Google Drive integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Production server
gunicorn==21.2.0
